# docs/specs/openapi.yaml
openapi: 3.0.0
info:
  title: DoggieGrowl API
  version: 1.0.0
  description: API for the DoggieGrowl application
servers:
  - url: http://localhost:3001/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: url
      required:
        - username
        - email
...
    Post:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        author:
          type: string
          format: uuid
      required:
        - title
        - content
        - author
    Comment:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        content:
          type: string
        author:
          type: string
          format: uuid
        post:
          type: string
          format: uuid
      required:
        - content
        - author
        - post
...
paths:
...
  /users/{id}:
    get:
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: User ID
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update user by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
...
  /forum:
    get:
      summary: Get all forum posts
      responses:
        '200':
          description: List of forum posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new forum post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: Post created successfully
  /forum/{id}:
    get:
      summary: Get forum post by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Post ID
      responses:
        '200':
          description: Post data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    put:
      summary: Update forum post by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Post ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '200':
          description: Post updated successfully
    delete:
      summary: Delete forum post by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Post ID
      responses:
        '200':
          description: Post deleted successfully
  /forum/{id}/comments:
    post:
      summary: Add a comment to a forum post
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Post ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '201':
          description: Comment added successfully
...